R Script Generator
Added hotkey for generating an R script that displays existing records in nice graphs. Errors in retrieving records now result in duplicate entries instead of null, so that a record may be taken at the end of the run. Singleplayer testing hook is now a hotkey instead of a line of code that needed to be commented in and out.