Analyser & timeStillUnsafe
First instance of the Analyser implemented, along with the framework of the Award class.

Fixed and modified timeStillPreTP to become timeStillUnsafe. Increments whenever the player is considered to be 'unsafe': not post tp, not post void fields, and not post void locus

Both timers also no longer increment when the run stopwatch is paused, to be consistent with the game's base stats